pyproject.toml
[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[project]
name = "ingest-relay"
version = "0.1.0"
description = "GitOps-managed sync bridge for on-prem connectors to GCS and Gemini Enterprise"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.11"
authors = [
  {name = "IngestRelay Contributors"}
]
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.6",
  "pydantic>=2.9.2",
  "pydantic-settings>=2.5.2",
  "PyYAML>=6.0.2",
  "SQLAlchemy>=2.0.36",
  "psycopg2-binary>=2.9.10",
  "pymysql>=1.1.1",
  "pymssql>=2.3.1",
  "oracledb>=2.5.1",
  "httpx>=0.27.2",
  "jinja2>=3.1.4",
  "google-cloud-storage>=2.19.0",
  "google-auth>=2.35.0",
  "tenacity>=9.0.0",
  "python-json-logger>=2.0.7",
  "typer>=0.12.5",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.3",
  "pytest-asyncio>=0.24.0",
  "pytest-cov>=5.0.0",
  "jsonschema>=4.23.0",
  "ruff>=0.7.1",
  "diff-cover>=9.4.1",
  "pip-audit>=2.9.0",
]

[project.scripts]
ingest-relay = "ingest_relay.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["ingest_relay"]
include = [
  "ingest_relay/templates/**",
  "ingest_relay/static/**",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
